Generic Popup with using MUI Dialog
