+++ psagot-client/src/components/GenericPopup.jsx
@@ -1,71 +1,57 @@
-import React, { useEffect, useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { fetchAllUserTypes } from "../features/userType/userTypeActions";
-import { Typography, Button, Container } from "@mui/material";
-import GenericPopup from "../components/GenericPopup";
+import React from "react";
+import PropTypes from "prop-types";
+import Button from "@mui/material/Button";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogContentText from "@mui/material/DialogContentText";
+import DialogTitle from "@mui/material/DialogTitle";
+import IconButton from "@mui/material/IconButton";
+import CloseIcon from "@mui/icons-material/Close";
 
-const HomePage = () => {
-  const dispatch = useDispatch();
-  const { userTypes, status, error } = useSelector((state) => state.userType);
-
-  useEffect(() => {
-    if (status === 'idle') {
-      dispatch(fetchAllUserTypes());
-    }
-  }, [status, dispatch]);
-
-  const [open, setOpen] = useState(false);
-  const [showButtons, setShowButtons] = useState(true); // מצב האם להראות כפתורים
-
-  const handleOpenPopup = () => setOpen(true);
-  const handleClosePopup = () => setOpen(false);
-
-  const handleConfirm = () => {
-    alert("כפתור אישור נלחץ");
-    setOpen(false);
-  };
-
-  const handleCancel = () => {
-    // alert("כפתור ביטול נלחץ");
-    setOpen(false);
-  };
-
-  const handleClickButton = () => {
-    alert("handle click button - userTypes" + JSON.stringify(userTypes));
-  };
-
-  const toggleButtons = () => {
-    setShowButtons(prevState => !prevState); // משנה את המצב של כפתורים
-  };
-
-  if (status === 'loading') return <Typography>Loading...</Typography>;
-  if (status === 'failed') return <Typography>Error: {error}</Typography>;
-
+const GenericPopup = ({ open, onClose, title, children, onConfirm, onCancel, showConfirmCancelButtons = true }) => {
   return (
-    <Container item style={{ textAlign: 'center', padding: 10 }}>
-      <Typography variant="h5">😀hello psagot project😀</Typography>
-      <Button onClick={handleClickButton}>Example of a function structure</Button>
-      <div>
-        {/* כפתור לפתיחת הפופ-אפ */}
-        <button onClick={handleOpenPopup}>פתח פופ-אפ</button>
-        
-        {/* כפתור לשנות את מצב הצגת כפתורי אישור וביטול */}
-        <button onClick={toggleButtons}>שנה מצב כפתורי אישור/ביטול</button>
-
-        {/* שימוש בקומפוננטה של GenericPopup */}
-        <GenericPopup
-          open={open}
-          onClose={handleClosePopup}
-          title="כותרת הפופ-אפ"
-          onConfirm={handleConfirm}
-          onCancel={handleCancel}
-          showConfirmCancelButtons={showButtons} 
-        >
-          זהו תוכן הפופ-אפ!
-        </GenericPopup>
-      </div>
-    </Container>
+    <Dialog
+      open={open}
+      onClose={onClose}
+      aria-labelledby="dialog-title">
+      <DialogTitle id="dialog-title">
+        {title}
+        <IconButton
+          aria-label="close"
+          onClick={onClose}>
+          <CloseIcon />
+        </IconButton>
+      </DialogTitle>
+      <DialogContent>
+        <DialogContentText>{children}</DialogContentText>
+      </DialogContent>
+      {showConfirmCancelButtons && (
+        <DialogActions>
+          {onCancel && (
+            <Button onClick={onCancel}>
+              ביטול
+            </Button>
+          )}
+          {onConfirm && (
+            <Button onClick={onConfirm} autoFocus>
+              אישור
+            </Button>
+          )}
+        </DialogActions>
+      )}
+    </Dialog>
   );
 };
 
-export default HomePage;
+GenericPopup.propTypes = {
+  open: PropTypes.bool.isRequired,
+  onClose: PropTypes.func.isRequired,
+  title: PropTypes.string,
+  children: PropTypes.node,
+  onConfirm: PropTypes.func,
+  onCancel: PropTypes.func,
+  showConfirmCancelButtons: PropTypes.bool,
+};
+
+export default GenericPopup;
