--- psagot-client/src/styles/GenericPopup.css
@@ -1,47 +1,0 @@
-/* עיצוב כללי של הפופ-אפ */
-.popup-container {
-    position: fixed;
-    top: 20%;
-    left: 50%;
-    transform: translate(-50%, -20%);
-    background: white;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    width: 300px;
-    text-align: center;
-}
-
-/* כפתור סגירה (X) בפינה השמאלית העליונה */
-.popup-close-button {
-    position: absolute;
-    top: 10px;
-    left: 10px;
-    background: transparent;
-    border: none;
-    font-size: 18px;
-    cursor: pointer;
-}
-
-/* עיצוב כפתורי אישור וביטול */
-.popup-buttons {
-    margin-top: 20px;
-    display: flex;
-    justify-content: space-around;
-}
-
-.popup-button {
-    padding: 8px 16px;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    color: white;
-}
-
-.popup-button.confirm {
-    background: green;
-}
-
-.popup-button.cancel {
-    background: red;
-}
